added validation checks for products and carts

+++ src/routes/carts.routes.js
@@ -1,6 +1,7 @@
 import { Router } from "express";
 import { CartManager } from "../dao/CartManager.js";
 import { isValidObjectId } from "mongoose";
+import { ProductManager } from "../dao/ProductManager.js";
 
 export const cartsRouter = Router();
 
@@ -30,6 +31,7 @@ cartsRouter.get("/:cid", async (req, res) => {
 
     try {
         const products = await CartManager.getCartProducts(cid);
+        if (!products) return res.status(404).json({ error: "El carrito no existe" });
         res.setHeader("Content-Type", "application/json");
         res.status(200).json(products);
     } catch (error) {
@@ -55,6 +57,14 @@ cartsRouter.post("/:cid/products/:pid", async (req, res) => {
         return res.status(400).json({ message: "Cid o Pid invalido" });
     }
     try {
+        const prodExist = await ProductManager.getProductsById(pid);
+        if (!prodExist) {
+            return res.status(404).json({error: `El producto con id ${pid} no existe`});
+        }
+
+        const cardExist = await CartManager.getCartProducts(cid);
+        if (!cardExist) return res.status(404).json({error: `El carrito con id ${cid} no existe`});
+
         await CartManager.addProductToCart(cid, pid);
         res.setHeader("Content-Type", "application/json");
         res.status(200).json({ message: "Producto agregado correctamente" });
@@ -81,9 +91,12 @@ cartsRouter.delete("/:cid/products/:pid", async (req, res) => {
         return res.status(400).json({ message: "Cid o Pid invalido" });
     }
     try {
+        const cardExist = await CartManager.getCartProducts(cid);
+        if (!cardExist) return res.status(404).json({error: `El carrito con id ${cid} no existe`});
+
         await CartManager.deleteProductFromCart(cid, pid);
         res.setHeader("Content-Type", "application/json");
-        res.status(200).json({ message: "Producto eliminado correctamente" });
+        return res.status(200).json({ message: "Producto eliminado correctamente" });
     } catch (error) {
         console.log(error);
         res.setHeader("Content-Type", "application/json");
@@ -107,6 +120,9 @@ cartsRouter.delete("/:cid", async (req, res) => {
         return res.status(400).json({ message: "Id invalido" });
     }
     try {
+        const cardExist = await CartManager.getCartProducts(cid);
+        if (!cardExist) return res.status(404).json({error: `El carrito con id ${cid} no existe`});
+
         await CartManager.deleteAllProducts(cid);
         res.setHeader("Content-Type", "application/json");
         res.status(200).json({ message: "Productos eliminados correctamente" });
@@ -122,7 +138,7 @@ cartsRouter.delete("/:cid", async (req, res) => {
 
 
 cartsRouter.put("/:cid", async(req, res)=> {
-    const {cid} = req.params;
+    const { cid } = req.params;
     const products = req.body;
 
     if (!cid )
@@ -134,8 +150,23 @@ cartsRouter.put("/:cid", async(req, res)=> {
         res.setHeader("Content-Type", "application/json");
         return res.status(400).json({ message: "Id invalido" });
     }
+
     if (!products) return res.status(400).json({error: "No se paso ningun producto"});
+
     try {
+        for (const prod of products) {
+            if (!isValidObjectId(prod.product)) {
+                return res.status(400).json({error: `El id del producto ${prod.product} no es valido`});
+            }
+            const prodExist = await ProductManager.getProductsById(prod.product);
+            if (!prodExist) {
+                return res.status(404).json({error: `El producto con id ${prod.product} no existe`});
+            }
+        }
+
+        const cardExist = await CartManager.getCartProducts(cid);
+        if (!cardExist) return res.status(404).json({error: `El carrito con id ${cid} no existe`});
+
         await CartManager.updateAllCart(cid, products);
         res.setHeader("Content-Type", "application/json");
         res.status(200).json({ message: "Carrito actualizado con los productos correctamente" });
@@ -169,6 +200,13 @@ cartsRouter.put("/:cid/products/:pid", async (req, res) => {
     if (isNaN(quantityNumber)) return res.status(400).json({error: "La cantidad debe ser un numero"});
 
     try {
+        const prodExist = await ProductManager.getProductsById(pid);
+        if (!prodExist) {
+            return res.status(404).json({error: `El producto con id ${pid} no existe`});
+        }
+        const cardExist = await CartManager.getCartProducts(cid);
+        if (!cardExist) return res.status(404).json({error: `El carrito con id ${cid} no existe`});
+
         await CartManager.updateProductQuantity(cid, pid, quantityNumber);
         res.setHeader("Content-Type", "application/json");
         res.status(200).json({ message: "Cantidad actualizada correctamente" });
